+++ lib/bean/res_options.dart
@@ -5,11 +5,13 @@ class ResOptions {
   int? statusCode;
   DateTime? responseTime; //ms
   int? duration; //ms
+  Map<String, List<String>>? headers;
   ResOptions({
     this.id,
     this.data,
     this.statusCode,
     this.responseTime,
     this.duration,
+    this.headers,
   });
 }